Update to use int64_t

+++ include/hgraph/builders/node_builder.h
@@ -18,7 +18,7 @@ namespace hgraph
                     std::optional<output_builder_ptr> error_builder_            = std::nullopt,
                     std::optional<output_builder_ptr> recordable_state_builder_ = std::nullopt);
 
-        virtual node_ptr make_instance(const std::vector<int64_t> &owning_graph_id, int node_ndx) const = 0;
+        virtual node_ptr make_instance(const std::vector<int64_t> &owning_graph_id, int64_t node_ndx) const = 0;
 
         virtual void release_instance(node_ptr &item) const {};
 
@@ -47,7 +47,7 @@ namespace hgraph
                           std::optional<output_builder_ptr> recordable_state_builder_, nb::callable eval_fn, nb::callable start_fn,
                           nb::callable stop_fn);
 
-        node_ptr make_instance(const std::vector<int64_t> &owning_graph_id, int node_ndx) const override;
+        node_ptr make_instance(const std::vector<int64_t> &owning_graph_id, int64_t node_ndx) const override;
 
         nb::callable eval_fn;
         nb::callable start_fn;
@@ -61,7 +61,7 @@ namespace hgraph
                                    std::optional<output_builder_ptr> output_builder_,
                                    std::optional<output_builder_ptr> error_builder_, nb::callable eval_fn);
 
-        node_ptr make_instance(const std::vector<int64_t> &owning_graph_id, int node_ndx) const override;
+        node_ptr make_instance(const std::vector<int64_t> &owning_graph_id, int64_t node_ndx) const override;
 
         nb::callable eval_fn;
     };
@@ -69,17 +69,17 @@ namespace hgraph
     struct BaseTsdMapNodeBuilder : BaseNodeBuilder
     {
         BaseTsdMapNodeBuilder(node_signature_ptr signature_, nb::dict scalars_,
-                              std::optional<input_builder_ptr>            input_builder_            = std::nullopt,
-                              std::optional<output_builder_ptr>           output_builder_           = std::nullopt,
-                              std::optional<output_builder_ptr>           error_builder_            = std::nullopt,
-                              std::optional<output_builder_ptr>           recordable_state_builder_ = std::nullopt,
-                              graph_builder_ptr                           nested_graph_builder      = {},
-                              const std::unordered_map<std::string, int> &input_node_ids = {}, int output_node_id = -1,
+                              std::optional<input_builder_ptr>                input_builder_            = std::nullopt,
+                              std::optional<output_builder_ptr>               output_builder_           = std::nullopt,
+                              std::optional<output_builder_ptr>               error_builder_            = std::nullopt,
+                              std::optional<output_builder_ptr>               recordable_state_builder_ = std::nullopt,
+                              graph_builder_ptr                               nested_graph_builder      = {},
+                              const std::unordered_map<std::string, int64_t> &input_node_ids = {}, int64_t output_node_id = -1,
                               const std::unordered_set<std::string> &multiplexed_args = {}, const std::string &key_arg = {});
 
         graph_builder_ptr                          nested_graph_builder;
-        const std::unordered_map<std::string, int> input_node_ids;
-        int                                        output_node_id;
+        const std::unordered_map<std::string, int64_t> input_node_ids;
+        int64_t                                        output_node_id;
         const std::unordered_set<std::string>      multiplexed_args;
         const std::string                          key_arg;
     };
@@ -88,7 +88,7 @@ namespace hgraph
     {
         using BaseTsdMapNodeBuilder::BaseTsdMapNodeBuilder;
 
-        node_ptr make_instance(const std::vector<int64_t> &owning_graph_id, int node_ndx) const override;
+        node_ptr make_instance(const std::vector<int64_t> &owning_graph_id, int64_t node_ndx) const override;
     };
 
 }  // namespace hgraph
